requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
influxdb==5.3.1
influxdb-client==1.37.0
msgpack==1.0.7
paho-mqtt==1.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
reactivex==4.0.4
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.6
